数量更新ページをコンポーネントに分離
